refactor(questionnaire): move Navbar component to root directory
The Navbar component was relocated from a nested directory to the root directory of the questionnaire module to simplify the project structure and improve maintainability
